HAAR-1937: refactor POST '/clients/{clientId}/duplicate' to 'base-clients/{baseClientId}/clients'
