--- src/main/kotlin/uk/gov/justice/digital/hmpps/authorizationserver/resource/ClientCredentialsController.kt
@@ -1,78 +1,0 @@
-package uk.gov.justice.digital.hmpps.authorizationserver.resource
-
-import com.microsoft.applicationinsights.TelemetryClient
-import jakarta.validation.constraints.NotBlank
-import jakarta.validation.constraints.Size
-import org.springframework.core.convert.ConversionService
-import org.springframework.http.HttpStatus
-import org.springframework.http.ResponseEntity
-import org.springframework.security.access.prepost.PreAuthorize
-import org.springframework.stereotype.Controller
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.ResponseStatus
-import uk.gov.justice.digital.hmpps.authorizationserver.config.AuthenticationFacade
-import uk.gov.justice.digital.hmpps.authorizationserver.service.ClientsService
-
-@Controller
-class ClientCredentialsController(
-  private val clientsService: ClientsService,
-  private val conversionService: ConversionService,
-  private val telemetryClient: TelemetryClient,
-  private val authenticationFacade: AuthenticationFacade,
-) {
-
-  @GetMapping("base-clients/{baseClientId}")
-  @ResponseStatus(HttpStatus.OK)
-  @PreAuthorize("hasRole('ROLE_OAUTH_ADMIN')")
-  fun viewClient(@PathVariable baseClientId: String): ResponseEntity<Any> {
-    return ResponseEntity.ok(
-      conversionService.convert(
-        clientsService.retrieveClientFullDetails(baseClientId),
-        ClientViewResponse::class.java,
-      ),
-    )
-  }
-}
-
-data class ClientViewResponse(
-  val clientId: String,
-  val scopes: List<String>,
-  val authorities: List<String>?,
-  val ips: List<String>?,
-  val jiraNumber: String?,
-  val databaseUserName: String?,
-  val validDays: Long?,
-  val accessTokenValidityMinutes: Long?,
-)
-
-data class ClientUpdateRequest(
-  val scopes: List<String>,
-  val authorities: List<String>,
-  val ips: List<String>,
-  val jiraNumber: String?,
-  val databaseUserName: String?,
-  val validDays: Long?,
-  val accessTokenValidityMinutes: Long?,
-)
-
-data class ClientRegistrationRequest(
-  @field:NotBlank(message = "clientId must not be blank")
-  @field:Size(max = 100, message = "clientId max size is 100")
-  val clientId: String?,
-
-  val scopes: List<String>?,
-  val authorities: List<String>?,
-  val ips: List<String>?,
-  val jiraNumber: String?,
-  val databaseUserName: String?,
-  val validDays: Long?,
-  val accessTokenValidityMinutes: Long?,
-)
-
-data class ClientRegistrationResponse(
-  val clientId: String,
-  val clientSecret: String,
-  val base64ClientId: String,
-  val base64ClientSecret: String,
-)
