+++ src/main/resources/db/dev/data/V900_1__authorization_consent.sql
@@ -1,2 +1,3 @@
 INSERT INTO oauth2_authorization_consent(registered_client_id, principal_name, authorities) VALUES
- ('ceab6c18-081b-44e9-8130-3b37a254108f', 'test-client-id', 'ROLE_OAUTH_ADMIN,ROLE_AUDIT,ROLE_TESTING');
+ ('ceab6c18-081b-44e9-8130-3b37a254108f', 'test-client-id', 'ROLE_OAUTH_ADMIN,ROLE_AUDIT,ROLE_TESTING'),
+ ('ceab6c18-081b-44e9-8130-3b37a254108g', 'test-duplicate-id', 'ROLE_OAUTH_ADMIN,ROLE_AUDIT,ROLE_TESTING');
